requirements.txt
# Basic requirements
bson>=0.5,<0.6

# Extensions
boto>=2.49,<3.0
google-cloud-storage>1.36,<=2.0

# Testing
pytest>=6.0.2,<6.1
coverage>4.5,<4.6  # downgraded on purpose for --concurrency flag

# Development
ipdb>=0.13
